Major commit

+++ routes/userRoutes.js
@@ -1,40 +1,34 @@
 const express = require('express');
-const { pgQuery } = require('../db.js');
+const { pgQuery } = require('../utils/db.js');
 
 const router = express.Router()
 
-router.get('/hello', (req, res) => {
-    res.json({message: "Hallo Welt!"})
-});
-
-router.get('/allUsers', async (req, res) => {
+router.get('/', async (req, res) => {
     try {
         const sql = "SELECT * FROM users;"
         const queryRes = await pgQuery(sql);
         return res.json(queryRes);
     } catch (err) {
         console.error("Error fetching users: " + err);
-        res.status(500).json( { error: `Internal server error: ${err}`} );
+        res.status(500).json( { message: `Internal server error: ${err}`} );
     }
 });
 
 router.post('/addUser', async (req, res) => {
     try {
         const userReq = req.body;
-        const userAlreadyExists = await userExists(userReq.email);
+        const userData = await userExists(userReq.email);
 
-        let sql = 'INSERT INTO users (name, email, password) VALUES ($1, $2, $3)';
-        let queryValues = [userReq.name, userReq.email, userReq.password];
+        let sql = 'INSERT INTO users (name, email, password, user_type) VALUES ($1, $2, $3, $4)';
+        let queryValues = [userReq.name, userReq.email, userReq.password, userReq.user_type];
+        if (userData.length > 0) return res.status(400).json({ message: "UNAUTHORIZED: User already exists." });
+            
+        await pgQuery(sql, queryValues);
 
-        if (!userAlreadyExists) {
-            await pgQuery(sql, queryValues);
-    
-            return res.status(200).json("User created successfuly.");
-        } else
-            throw new Error("User already exists.");
+        return res.status(200).json({ message: "User created successfuly." });
     } catch (err) {
         console.error(err);
-        res.status(500).json( { error: `Internal server error: ${err.message}`} );
+        res.status(500).json( { message: `Internal server error: ${err.message}`} );
     }
 });
 
@@ -42,25 +36,58 @@ router.delete('/removeUser', async (req, res) => {
     try {
         const userReq = req.body;
 
+        const userData = await userExists(userReq.email);
+
+        // If user does NOT exists, returns
+        if (!userData) return res.status(500).json({ message: `User \"${userReq.email}\" does not exists in database` });
+        
+        // If user has dependencies and haven't forceDeleted, returns
+        const hasDependencies = await hasUploads(userData.id);
+        if (hasDependencies && !userReq.forceDelete) return res.status(403).json({ message: "FORBIDDEN: Cannot delete user with uploads. Set forceDelete to true to override." });
+        
+        // If user forceDeletes the dependencies, continue
         const params = [userReq.email];
         const sql = "DELETE FROM users WHERE email = $1;"
         await pgQuery(sql, params);
 
-        return res.status(200).json("User \"" + userReq.email + "\" deleted successfully")
+        return res.status(200).json({ message: "User \"" + userReq.email + "\" deleted successfully" })
     } catch (err) {
-        res.status(500).json("Internal server error: " + err.message);
+        res.status(500).json({ message: "Internal server error: " + err.message });
     }
 });
 
 async function userExists(email) {
+    /*
+    * Checks if user exists in database
+    * Returns an array with [userId, true] if exists, [null, false] otherwise
+    */
     try {
         let sql = 'SELECT * FROM users WHERE email = $1;';
         let queryValues = [email];
         
         const user = await pgQuery(sql, queryValues);
-        return user.length > 0; // Returns true if exists
+        return user; // Returns true if exists
     } catch (err) {
-        console.error(`Error in getUser function: ${err}`);
+        throw err;
+    }
+}
+
+async function hasUploads(id) {
+    try {
+        const params = [id];
+        
+        let sql = "SELECT * FROM user_orders WHERE user_id = $1;";
+        const user_orders = await pgQuery(sql, params);
+
+        sql = "SELECT * FROM images WHERE uploader_id = $1;"
+        const images = await pgQuery(sql, params);
+
+        sql = "SELECT * FROM orders WHERE user_id = $1;"
+        const orders = await pgQuery(sql, params);
+
+        // If has relation, returns true
+        return (user_orders.length > 0 || images.length > 0 || orders.length > 0); 
+    } catch (err) {
         throw err;
     }
 }
